# add generic GCC and Clang compiler flags for RISC-V