This adds compiler flags (GCC and Clang) for generic builds on RISC-V. It's a bit more involved than on for instance x86_64, because of all the different extensions and ABIs, see:
https://gcc.gnu.org/onlinedocs/gcc/RISC-V-Options.html#index-march-14
https://five-embeddev.com/toolchain/2019/06/26/gcc-targets/

I've picked a sensible default, which also is in line with what Gentoo uses:
https://wiki.gentoo.org/wiki/RISC-V_Multilib

The optimal flags are a bit more complex, since there is no such things as `-march=native` for RISC-V. I'll try to take a stab at in a follow-up PR.