@bedroge 2024-03-29
Before:
```
== FAILED: Installation ended unsuccessfully (build directory: /nvme/eb/build/zlib/1.2.13/GCCcore-13.2.0): build failed (first 300 chars): Don't know how to set generic optarch flags for RISC-V-64/RISC-V! Use --optarch='<flags>' to override (see http://easybuild.readthedocs.io/en/latest/Controlling_compiler_optimization_flags.html for details) and consider contributing your settings back (see http://easybuild.readthedocs.io/en/latest/C (took 1 secs)
```

After:
```
user@starfive:/nvme/gcc$ eb zlib-1.2.13-GCCcore-13.2.0.eb --installpath=/nvme/easybuild-generic --optarch=GENERIC
== Temporary log file in case of crash /nvme/eb/tmp/eb-4vr3683t/easybuild-uleen25o.log
== found valid index for /nvme/easybuild/software/EasyBuild/4.9.0/easybuild/easyconfigs, so using it...
== found valid index for /nvme/easybuild/software/EasyBuild/4.9.0/easybuild/easyconfigs, so using it...
== processing EasyBuild easyconfig /nvme/easybuild/software/EasyBuild/4.9.0/easybuild/easyconfigs/z/zlib/zlib-1.2.13-GCCcore-13.2.0.eb
== building and installing zlib/1.2.13-GCCcore-13.2.0...
== fetching files...
== creating build dir, resetting environment...
== unpacking...
== patching...
== preparing...
== configuring...
== ... (took 1 secs)
== building...
== ... (took 9 secs)
== testing...
== installing...
== taking care of extensions...
== restore after iterating...
== postprocessing...
== sanity checking...
== cleaning up...
== creating module...
== ... (took 1 secs)
== permissions...
== packaging...
== COMPLETED: Installation ended successfully (took 15 secs)
== Results of the build can be found in the log file(s) /nvme/easybuild-generic/software/zlib/1.2.13-GCCcore-13.2.0/easybuild/easybuild-zlib-1.2.13-20240329.081717.log
== Build succeeded for 1 out of 1
== Temporary log file(s) /nvme/eb/tmp/eb-4vr3683t/easybuild-uleen25o.log* have been removed.
== Temporary directory /nvme/eb/tmp/eb-4vr3683t has been removed.

# some relevant lines from the build log:
gcc -O2 -ftree-vectorize -march=rv64gc -mabi=lp64d -fno-math-errno -fPIC -D_LARGEFILE64_SOURCE=1 -DHAVE_HIDDEN  -c -o gzlib.o gzlib.c
gcc -O2 -ftree-vectorize -march=rv64gc -mabi=lp64d -fno-math-errno -fPIC -D_LARGEFILE64_SOURCE=1 -DHAVE_HIDDEN  -c -o gzread.o gzread.c
```